% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesaran_smith.R
\name{pesaran_smith}
\alias{pesaran_smith}
\title{Teste de Significância para Impacto Médio (Pesaran & Smith) - Horizonte Fixo}
\usage{
pesaran_smith(model, target_var, y_hat, y)
}
\arguments{
\item{model}{Objeto da classe 'varest'.}

\item{target_var}{String com o nome da variável de interesse.}

\item{y_hat}{Objeto com as previsões.}

\item{y}{Objeto com os dados reais.}
}
\value{
Lista com impacto, estatística T e p-valor.
}
\description{
Teste de Significância para Impacto Médio (Pesaran & Smith) - Horizonte Fixo
}
