% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesaran_smith_cum.R
\name{pesaran_smith_cum}
\alias{pesaran_smith_cum}
\title{Teste de Significância para Impacto Médio Cumulativo (Pesaran & Smith)}
\usage{
pesaran_smith_cum(model, target_var, y_hat, y)
}
\arguments{
\item{model}{Objeto da classe 'varest'.}

\item{target_var}{String com o nome da variável de interesse.}

\item{y_hat}{Vetor completo de previsões.}

\item{y}{Vetor completo de dados reais.}
}
\value{
Um data.frame contendo o impacto, estatística T e p-valor para cada passo h.
}
\description{
Teste de Significância para Impacto Médio Cumulativo (Pesaran & Smith)
}
