% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pesaran_smith_cum.R
\name{plot_pesaran_smith_cum}
\alias{plot_pesaran_smith_cum}
\title{Plot do Impacto Médio Cumulativo}
\usage{
plot_pesaran_smith_cum(
  df_cum,
  title = "Impacto Médio Cumulativo (Pesaran & Smith)"
)
}
\arguments{
\item{df_cum}{Dataframe retornado por calc_pesaran_smith_cum.}

\item{title}{Título opcional para o gráfico.}
}
\value{
Um objeto ggplot2.
}
\description{
Plot do Impacto Médio Cumulativo
}
